Full comments, Interface updates
Interface is nearly fully functional and commented.

Things to do still, search isn't working yet however it will return
results in a similar fashion to findUserListing and findBookListing so
you can temporarily use that function.

Listing is functional. If we want to make them deletable rather than
just not available that will need to be added.